name typo fix in  namespace
