+++ src/MithgardWpf.App/Pages/Home/HomePageView.xaml
@@ -3,7 +3,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:aap="clr-namespace:MithgardWpf.App.Core.Animations.AttachedProperties"
-    xmlns:cv="clr-namespace:MithgardWpf.App.Common.ValueConvertors"
+    xmlns:cv="clr-namespace:MithgardWpf.App.Common.ValueConverters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="clr-namespace:MithgardWpf.App.Pages.Home"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
